+++ apps/ocpp/consumers/base/notifications.py
@@ -1,8 +1,8 @@
-"""Notification handlers for status, diagnostics, firmware, and security.
+"""Notification handlers for diagnostics, log, and security events.
 
 These adapters group OCPP 1.6/2.x notification actions whose legacy handlers
-perform DB side effects such as firmware deployment state updates, log request
-updates, and security event persistence.
+perform DB side effects such as log request updates and security event
+persistence.
 """
 
 from typing import Protocol
@@ -11,19 +11,11 @@ from apps.ocpp.payload_types import HandlerPayload, HandlerResponse
 
 
 class NotificationConsumer(Protocol):
-    async def _handle_publish_firmware_status_notification_action_legacy(
-        self, payload: HandlerPayload, msg_id: str, raw: str | None, text_data: str | None
-    ) -> HandlerResponse: ...
-
     async def _handle_diagnostics_status_notification_action_legacy(
         self, payload: HandlerPayload, msg_id: str, raw: str | None, text_data: str | None
     ) -> HandlerResponse: ...
 
     async def _handle_log_status_notification_action_legacy(
-        self, payload: HandlerPayload, msg_id: str, raw: str | None, text_data: str | None
-    ) -> HandlerResponse: ...
-
-    async def _handle_firmware_status_notification_action_legacy(
         self, payload: HandlerPayload, msg_id: str, raw: str | None, text_data: str | None
     ) -> HandlerResponse: ...
 
@@ -37,15 +29,6 @@ class NotificationHandler:
 
     def __init__(self, consumer: NotificationConsumer) -> None:
         self.consumer = consumer
-
-    async def handle_publish_firmware_status(
-        self, payload: HandlerPayload, msg_id: str, raw: str | None, text_data: str | None
-    ) -> HandlerResponse:
-        """Handle OCPP 2.x firmware publish notifications with DB updates."""
-
-        return await self.consumer._handle_publish_firmware_status_notification_action_legacy(
-            payload, msg_id, raw, text_data
-        )
 
     async def handle_diagnostics_status(
         self, payload: HandlerPayload, msg_id: str, raw: str | None, text_data: str | None
@@ -62,15 +45,6 @@ class NotificationHandler:
         """Handle log status notifications and persist log delivery progress."""
 
         return await self.consumer._handle_log_status_notification_action_legacy(
-            payload, msg_id, raw, text_data
-        )
-
-    async def handle_firmware_status(
-        self, payload: HandlerPayload, msg_id: str, raw: str | None, text_data: str | None
-    ) -> HandlerResponse:
-        """Handle OCPP 1.6 firmware status notifications with deployment writes."""
-
-        return await self.consumer._handle_firmware_status_notification_action_legacy(
             payload, msg_id, raw, text_data
         )
